pim/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
fastapi==0.115.12
h11==0.14.0
httptools==0.6.4
grpcio==1.71.0
grpcio-tools==1.71.0
idna==3.10
pydantic==2.11.2
pydantic_core==2.33.1
python-dotenv==1.1.0
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.1
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0.1
